chore: jasmin config (display, migration bigint string error)

+++ migrations/sqls/20210630055210-orders-table-up.sql
@@ -2,5 +2,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
     status VARCHAR(15) DEFAULT 'active',
-    user_id bigint REFERENCES users(id)
+    user_id int REFERENCES users(id)
 );
